Begin switching dashboard to table view
This commit breaks part of the frontend; just a temporary commit to a
new branch for syncing purposes, this should hopefully be squashed away
and disappear from later commit history (if I can figure out how).

Implementation still partial.  TODO comments in the html template
indicate the major missing pieces.  Next step is to build or identify
the backend queries to support it.