Fix bug where other games bleed into form choices
On the game dashboard page, in response to invalid POST requests,
objects belonging to other games (besides the one being displayed on the
dashboard) would show up in the select choices in the ClueRelationForm.

This was caused by the view failing to set the form Field's queryset
attribute in the case of a POST request.

Fixed in a more DRY way, by moving the queryset filter out of the view
function, and into the __init__ function of the form itself.